+++ changelog.md
@@ -8,6 +8,7 @@
 - Bugfix: Deadlock on node down event
 - Bugfix: Skip compact operation when only one SSTable exists
 - Bugfix: Set failed state on future when server operation returns an error
+- Bugfix: Flush full memtables also on tuple delete operation
 
 ### Version 0.1.2 (Alpha) - 14.09.2016
 - Implemented the recovery service